refactor(chat): extract native view props
